+++ tests/Deedle.Tests.Console/Program.fs
@@ -74,7 +74,7 @@ let testAll () =
 open Deedle
 
 let rnd = System.Random(0)
-
+(*
 let s0 = series <| Array.init 10000 (fun i -> i => rnd.NextDouble())
 let f0 = frame [ for i in 0 .. 10 -> i => s0 ]
 let f1 = frame [ for i in 0 .. 100 -> i => s0 ]
@@ -84,8 +84,22 @@ let s2 = series <| Array.init 1000000 (fun i -> i*2+1 => rnd.NextDouble())
 
 let ss1 = series <| Array.init 100 (fun i -> i*2 => rnd.NextDouble())
 let ss2 = series <| Array.init 100 (fun i -> i*2+1 => rnd.NextDouble())
-
+*)
 let testOne() =      
+  // 970 ~~> 457
+  printfn "Creating lots of small series"
+  timed 10 (fun () ->
+    let vs = [ for i in 0 .. 10 -> i => float i ]
+    Array.init 100000 (fun _ -> series vs) |> ignore
+  )
+  // 196 ~~> 122
+  printfn "Creating few large series"
+  timed 10 (fun () ->
+    let vs = [ for i in 0 .. 10000 -> i => float i ]
+    Array.init 100 (fun _ -> series vs) |> ignore
+  )
+
+(*
   timed 5 (fun () ->
     s1.ZipInner(s1)
     |> ignore
@@ -94,6 +108,7 @@ let testOne() =
     s1.ZipInner(s1)
     |> ignore
   )
+*)
   // 1288ms
   // 1286ms